using a LTS because openssh-server in the apt rep
I tried using Artful Aardvark, but openssh-server is not listed in the
repositories. Eventhough adding repositories for previous releases of
Ubuntu would work, it is indicative of a larger problem -- packages not
being available as expected by the end-user. The end-user is expected to
install a few other tools as needed and needs a base that works out of
the box.